rename: types 패키지 정리
각 관련 도메인 패키지 내부로 이동
