+++ src/main/java/com/foodielog/server/feed/Feed.java
@@ -1,17 +1,30 @@
 package com.foodielog.server.feed;
 
-import com.foodielog.server.restaurant.Restaurant;
-import com.foodielog.server.types.ContentStatus;
-import com.foodielog.server.user.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import java.sql.Timestamp;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.*;
-import java.sql.Timestamp;
+import com.foodielog.server.feed.type.ContentStatus;
+import com.foodielog.server.restaurant.Restaurant;
+import com.foodielog.server.user.User;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
@@ -19,31 +32,31 @@ import java.sql.Timestamp;
 @Table(name = "feed_tb")
 @Entity
 public class Feed {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "restaurant_id")
-    private Restaurant restaurant;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "restaurant_id")
+	private Restaurant restaurant;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
 
-    @Column(nullable = false, length = 255)
-    private String thumbnailUrl;
+	@Column(nullable = false, length = 255)
+	private String thumbnailUrl;
 
-    @Column(nullable = false, length = 2000)
-    private String content;
+	@Column(nullable = false, length = 2000)
+	private String content;
 
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 7)
-    private ContentStatus status;
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false, length = 7)
+	private ContentStatus status;
 
-    @CreationTimestamp
-    private Timestamp createdAt;
+	@CreationTimestamp
+	private Timestamp createdAt;
 
-    @UpdateTimestamp
-    private Timestamp updatedAt;
+	@UpdateTimestamp
+	private Timestamp updatedAt;
 }
