+++ src/main/java/com/foodielog/server/notification/type/NotificationType.java
@@ -1,0 +1,12 @@
+package com.foodielog.server.notification.type;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum NotificationType {
+	FEED("피드"),
+	LIKE("좋아요"),
+	FOLLOW("팔로우");
+
+	private final String label;
+}
