+++ src/main/java/com/foodielog/server/reply/Reply.java
@@ -1,43 +1,56 @@
 package com.foodielog.server.reply;
 
-import com.foodielog.server.feed.Feed;
-import com.foodielog.server.types.ContentStatus;
-import com.foodielog.server.user.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import java.sql.Timestamp;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.*;
-import java.sql.Timestamp;
+import com.foodielog.server.feed.Feed;
+import com.foodielog.server.feed.type.ContentStatus;
+import com.foodielog.server.user.User;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Table(name = "reply_tb")
 @Entity
 public class Reply {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "feed_id")
-    private Feed feed;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "feed_id")
+	private Feed feed;
 
-    @Column(length = 150, nullable = false)
-    private String content;
+	@Column(length = 150, nullable = false)
+	private String content;
 
-    @Enumerated(EnumType.STRING)
-    private ContentStatus status;
+	@Enumerated(EnumType.STRING)
+	private ContentStatus status;
 
-    @CreationTimestamp
-    private Timestamp createdAt;
+	@CreationTimestamp
+	private Timestamp createdAt;
 
-    @UpdateTimestamp
-    private Timestamp updatedAt;
+	@UpdateTimestamp
+	private Timestamp updatedAt;
 }
