--- src/main/java/com/foodielog/server/types/ContentStatus.java
@@ -1,8 +1,0 @@
-package com.foodielog.server.types;
-
-public enum ContentStatus {
-    NORMAL,
-    HIDDEN,
-    DELETED
-    ;
-}
