--- src/main/java/com/foodielog/server/types/Flag.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.types;
-
-public enum Flag {
-    YES,
-    NO
-    ;
-}
