--- src/main/java/com/foodielog/server/types/ManagementStatus.java
@@ -1,8 +1,0 @@
-package com.foodielog.server.types;
-
-public enum ManagementStatus {
-    UNPROCESSED,
-    APPROVED,
-    REJECTED
-    ;
-}
