--- src/main/java/com/foodielog/server/types/ReportType.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.types;
-
-public enum ReportType {
-    FEED,
-    REPLY
-    ;
-}
