--- src/main/java/com/foodielog/server/types/Role.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.types;
-
-public enum Role {
-    USER,
-    ADMIN
-    ;
-}
