--- src/main/java/com/foodielog/server/types/UserStatus.java
@@ -1,8 +1,0 @@
-package com.foodielog.server.types;
-
-public enum UserStatus {
-    NORMAL,
-    WITHDRAW,
-    BLOCK
-    ;
-}
