+++ src/main/java/com/foodielog/server/user/type/BlockReason.java
@@ -1,0 +1,4 @@
+package com.foodielog.server.user.type;
+
+public enum BlockReason {
+}
