+++ src/main/java/com/foodielog/server/user/type/WithDrawReason.java
@@ -1,0 +1,5 @@
+package com.foodielog.server.user.type;
+
+public enum WithDrawReason {
+
+}
